+++ CHANGELOG.md
@@ -2,8 +2,18 @@
 
 ## [0.4.0.0] - 15 May 2023
 ### Changes
-- Fixed a bug in `PowerShell-Scraping\Modules\SeleniumWhatsApp.psm1::Set-WhatsAppMessageBarText` where messages with line breaks/carriage returns would be sent automatically.
-- Added `PowerShell-Scraping\Modules\SeleniumWhatsApp.psm1::Set-WhatsAppAttachment` to add attachments to the message, with an instant send or a lazy send option.
+- In `PowerShell-Scraping\Modules\SeleniumWhatsApp.psm1`...
+    - Added:
+        - `Set-WhatsAppAttachment`
+        - `Send-WhatsAppAttachment`
+        - `Send-WhatsAppMessage`
+        - `Select-WhatsAppElement`
+        - `Test-WhatsAppAccount`
+        - `Get-WhatsAppMessageHyperLinks`
+    - Renamed:
+        - `Set-WhatsAppMessageBarText` to `Set-WhatsAppMessage`
+    - Fixed:
+        - A bug in `Set-WhatsAppMessage` where messages with line breaks/carriage returns would be sent automatically.
 
 ## [0.3.0.0] - 11 May 2023
 ### Changes